+++ claude_discord/cogs/__init__.py
@@ -2,6 +2,7 @@
 
 from .auto_upgrade import AutoUpgradeCog
 from .claude_chat import ClaudeChatCog
+from .scheduler import SchedulerCog
 from .session_manage import SessionManageCog
 from .skill_command import SkillCommandCog
 from .webhook_trigger import WebhookTriggerCog
@@ -9,6 +10,7 @@ from .webhook_trigger import WebhookTriggerCog
 __all__ = [
     "AutoUpgradeCog",
     "ClaudeChatCog",
+    "SchedulerCog",
     "SessionManageCog",
     "SkillCommandCog",
     "WebhookTriggerCog",